requirements.txt
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
python-slugify==8.0.4
Jinja2==3.1.4
